Add the --map_binary_to_string compat flag to update-to-latest.sh.
I added this flag in https://chromium-review.googlesource.com/c/deps/inspector_protocol/+/1285695. It makes it so that protocol binary types in .pdl files are represented as string in the .json protocol description files. This means that it's up to the client code to handle the base64 decoding and encoding. On the wire, it's a JSON string either way.