Update the template to include feature requests
I've come here to file a feature request. I'm not surprised to see a no-discussion style issue template here, redirecting me to CRbug, but I find it odd that feature requests are not mentioned once in it. Are feature requests not accepted for CDP? Only bug reports? I filed my feature request with CRbug anyway, but it's really not clear what is the stance on those here, or there. The name is CRbug after all. Do I have to get hired by Google, befriend the people on the CDP team, work towards a transfer to the CDP team and then pitch my idea to the CDP PM?