Add top level `@public` doc comment.
In Puppeteer we're using this package as part of the source to generate
documentation for Puppeteer.

By having docs generated from the protocol's d.ts file we give users a
nice experience where a function that's documented as taking an argument
of `Protocol.X` is documented with a link to `Protocol.X`.

This works nicely out of the box as all of the protocol has some
documenting comments alongside it (we will also point users towards the
devtools protocol viewer), but one small change we need to satisfy the
API documentation tool we use
(https://api-extractor.com/pages/tsdoc/doc_comment_syntax/) is to add a
top level comment with a `@public` declaration.