Explicitly specify folders to be included in npm package
Previously, `.github/*` was included in the package. I’ve confirmed with `npm pack` that this is the only difference introduced by this patch.

Note that I’ve removed `.npmignore` since it’s a denylist (and thus easy to get out of date, which happened here), whereas `package.json#files` is an allowlist.